Working on the docs.
Created "generated.py" which generates the .rst files for the methods of
classes. Furthermore it changes the class.rst so that methods are
clickable inside the html-docs.

The makefile has a new structure therefore:
1. sphinx-make
2. generate.py
3. sphinx make

the second sphinx-make is for linking the method-rst files to the class
rst-files.

TODO: methods of parent classes are now in the docs as well, because
they are not secret (__secret_method).