Working on the docs.
Created "generated.py" which generates the .rst files for the methods of
classes. Furthermore it changes the class.rst so that methods are
clickable inside the html-docs.

The makefile has a new structure therefore:
1. sphinx-make
2. generate.py
3. sphinx make

the second sphinx-make is for linking the method-rst files to the class
rst-files.

Secret methods are not in the docs anymore aas well as methods from
parent classes, as long as they are at least private (_methods)

Function names are right now and not containing the whole path anymore.
So far the automatic documentation creation is working then.